[jif]: don't fail on adding an ordering chunk that is not mapped
In fact, if the recorded log is very extensive, it is possible that,
after restoring (and starting to record) a new region of memory is
mapped. After this the accesses to that region can still get recorded,
but will be meaningless when looking at the original JIF. As such, we
should just ignore these mappings.

We should discuss whether we should halt the sequence there.