jif: break ordering chunks by intervals when writing
there is a nice invariant that we want that jifs created by jiftool
should never have an ordering chunk that spans multiple intervals.

This is rarely the case because we usually build the itrees first and
then add ordering chunks (and then, with setup-prefetch, we can break
*intervals* by ordering chunks).

However, if a user first `add-ord`s and then `build-itrees`, the
ordering chunks might span multiple itrees.

This change enforces that ordering chunks are split by interval before
being written back to disk