jif: change how ordering works (format version -> 4)
This commit changes the way ordering works.

First, ordering chunks now have the (minimum) timestamp from the the trace that references it.

Second, ordering changes depending on whether prefetch is enabled or
not.

For non-prefetch workloads, both ordering and the data are ordered by
timestamp.

For prefetch workloads, the ordering section is ordered by address. This
enables the prefetcher to do a single VMA lookup per VMA. The data
section is ordered in two partitions. The first with intervals that are
written to. The second with intervals that are not. Internally, these
are also ordered by address.